+++ NEWS.md
@@ -8,6 +8,10 @@
 
 * `mlogit` (package **mlogit**)
 
+## General
+
+* `plot()` now can also create partial residuals plots. There, arguments `residuals`, `residuals.type` and `residuals.line` were added to add partial residuals, the type of residuals and a possible loess-fit regression line for the residual data.
+
 ## Bug fixes
 
 * The message for models with a back-transformation to the response scale (all non-Gaussian models), that standard errors are still on the link-scale, did not show up for models of class `glm` since some time. Should be fixed now.